+++ src/providers/strategies/helpers/signTransactions/helpers/getCommonData/helpers/tests/getGasPriceOptions.test.ts
@@ -1,6 +1,6 @@
 import { testAddress } from '__mocks__';
 import { IPlainTransactionObject } from 'lib/sdkCore';
-import { getPpuOptions } from '../getPpuOptions';
+import { getGasPriceOptions } from '../getGasPriceOptions';
 
 const secondTx: IPlainTransactionObject = {
   nonce: 0,
@@ -14,9 +14,9 @@ const secondTx: IPlainTransactionObject = {
   version: 1
 };
 
-describe('getPpuOptions', () => {
+describe('getGasPriceOptions test', () => {
   it('should return the correct gas price options', () => {
-    const ppuOptions = getPpuOptions({
+    const gasPriceOptions = getGasPriceOptions({
       shard: 1,
       gasStationMetadata: [
         {
@@ -36,10 +36,10 @@ describe('getPpuOptions', () => {
       initialGasPrice: 1_000_000_000
     });
 
-    expect(ppuOptions).toEqual([
-      { label: 'Standard', value: 0 },
-      { label: 'Fast', value: 11760000 },
-      { label: 'Faster', value: 19287760 }
+    expect(gasPriceOptions).toEqual([
+      { label: 'Standard', value: 1000000000 },
+      { label: 'Fast', value: 1065456733 },
+      { label: 'Faster', value: 1747472258 }
     ]);
   });
 });