+++ src/utils/decoders/base64Utils.ts
@@ -41,7 +41,7 @@ export function isStringBase64(str: string) {
       // it is a regular base64 string
       return true;
     }
-  } catch (e) {
+  } catch (_) {
     return false;
   }